feat: Add Service Worker and improve animations
Introduces a Service Worker for offline capabilities and enhances the UI with new animations for a smoother user experience.

The Service Worker registration is added to `index.tsx` to enable PWA features.
The `index.html` file is updated with a new font weight and Tailwind CSS configurations for the 'slide-left' animation.
The `types.ts` file includes a significant expansion of the `Anamnesis` interface to better structure patient history data, and new types for notifications and logging.
The `App.tsx` component now includes logic for Service Worker registration and imports new components and types.
The `Dashboard.tsx` component sees refactoring to include more summary data and UI elements, improving its role as an overview page.